Improve the adapters
